pyproject.toml
[project]
name = "mcp-server"
version = "0.0.1"
description = "Simple MCP server test with remote tools support."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "mcp==1.26.0",
  "pandas==3.0.1",
  "pyyaml==6.0.3",

]

[project.scripts]
mcp-server = "mcp_server.server:main"

[dependency-groups]
dev = [
  "pytest",
]

[build-system]
requires = ["uv_build>=0.10.8,<0.11.0"]
build-backend = "uv_build"

